+++ apps/client/src/app/components/admin-market-data/admin-market-data.component.ts
@@ -1,8 +1,8 @@
 import { UserService } from '@ghostfolio/client/services/user/user.service';
 import {
   DEFAULT_COLOR_SCHEME,
-  DEFAULT_PAGE_SIZE,
-  locale
+  DEFAULT_LOCALE,
+  DEFAULT_PAGE_SIZE
 } from '@ghostfolio/common/config';
 import {
   canDeleteAssetProfile,
@@ -10,11 +10,11 @@ import {
 } from '@ghostfolio/common/helper';
 import {
   AssetProfileIdentifier,
+  AssetProfileItem,
   Filter,
   InfoItem,
   User
 } from '@ghostfolio/common/interfaces';
-import { AdminMarketDataItem } from '@ghostfolio/common/interfaces/admin-market-data.interface';
 import { hasPermission, permissions } from '@ghostfolio/common/permissions';
 import { GfSymbolPipe } from '@ghostfolio/common/pipes';
 import { GfActivitiesFilterComponent } from '@ghostfolio/ui/activities-filter';
@@ -152,7 +152,7 @@ export class GfAdminMarketDataComponent implements AfterViewInit, OnInit {
     }
   ];
   protected readonly canDeleteAssetProfile = canDeleteAssetProfile;
-  protected dataSource = new MatTableDataSource<AdminMarketDataItem>();
+  protected dataSource = new MatTableDataSource<AssetProfileItem>();
   protected defaultDateFormat: string;
   protected readonly displayedColumns: string[] = [];
   protected readonly filters$ = new Subject<Filter[]>();
@@ -160,7 +160,7 @@ export class GfAdminMarketDataComponent implements AfterViewInit, OnInit {
   protected readonly isUUID = isUUID;
   protected pageSize = DEFAULT_PAGE_SIZE;
   protected placeholder = '';
-  protected readonly selection = new SelectionModel<AdminMarketDataItem>(true);
+  protected readonly selection = new SelectionModel<AssetProfileItem>(true);
   protected totalItems = 0;
   protected user: User;
 
@@ -375,8 +375,8 @@ export class GfAdminMarketDataComponent implements AfterViewInit, OnInit {
 
     this.selection.clear();
 
-    this.adminService
-      .fetchAdminMarketData({
+    this.dataService
+      .fetchAssetProfiles({
         sortColumn,
         sortDirection,
         filters: this.activeFilters,
@@ -384,15 +384,15 @@ export class GfAdminMarketDataComponent implements AfterViewInit, OnInit {
         take: this.pageSize
       })
       .pipe(takeUntilDestroyed(this.destroyRef))
-      .subscribe(({ count, marketData }) => {
+      .subscribe(({ assetProfiles, count }) => {
         this.totalItems = count;
 
         this.dataSource = new MatTableDataSource(
-          marketData.map((marketDataItem) => {
+          assetProfiles.map((assetProfile) => {
             return {
-              ...marketDataItem,
+              ...assetProfile,
               isBenchmark: this.benchmarks.some(({ id }) => {
-                return id === marketDataItem.id;
+                return id === assetProfile.id;
               })
             };
           })
@@ -429,7 +429,7 @@ export class GfAdminMarketDataComponent implements AfterViewInit, OnInit {
             colorScheme:
               this.user?.settings.colorScheme ?? DEFAULT_COLOR_SCHEME,
             deviceType: this.deviceType(),
-            locale: this.user?.settings?.locale ?? locale
+            locale: this.user?.settings?.locale ?? DEFAULT_LOCALE
           } satisfies AssetProfileDialogParams,
           height: this.deviceType() === 'mobile' ? '98vh' : '80vh',
           width: this.deviceType() === 'mobile' ? '100vw' : '50rem'
@@ -464,7 +464,7 @@ export class GfAdminMarketDataComponent implements AfterViewInit, OnInit {
           autoFocus: false,
           data: {
             deviceType: this.deviceType(),
-            locale: this.user?.settings?.locale ?? locale
+            locale: this.user?.settings?.locale ?? DEFAULT_LOCALE
           } satisfies CreateAssetProfileDialogParams,
           width: this.deviceType() === 'mobile' ? '100vw' : '50rem'
         });