+++ libs/common/src/lib/interfaces/index.ts
@@ -4,13 +4,10 @@ import type { Activity, ActivityError } from './activities.interface';
 import type { AdminData } from './admin-data.interface';
 import type { AdminJobs } from './admin-jobs.interface';
 import type { AdminMarketDataDetails } from './admin-market-data-details.interface';
-import type {
-  AdminMarketData,
-  AdminMarketDataItem
-} from './admin-market-data.interface';
 import type { AdminUser } from './admin-user.interface';
 import type { AssetClassSelectorOption } from './asset-class-selector-option.interface';
 import type { AssetProfileIdentifier } from './asset-profile-identifier.interface';
+import type { AssetProfileItem } from './asset-profile-item.interface';
 import type { BenchmarkProperty } from './benchmark-property.interface';
 import type { Benchmark } from './benchmark.interface';
 import type { Coupon } from './coupon.interface';
@@ -50,6 +47,7 @@ import type { AdminUsersResponse } from './responses/admin-users-response.interf
 import type { AiPromptResponse } from './responses/ai-prompt-response.interface';
 import type { AiServiceHealthResponse } from './responses/ai-service-health-response.interface';
 import type { ApiKeyResponse } from './responses/api-key-response.interface';
+import type { AssetProfilesResponse } from './responses/asset-profiles-response.interface';
 import type { AssetResponse } from './responses/asset-response.interface';
 import type { BenchmarkMarketDataDetailsResponse } from './responses/benchmark-market-data-details-response.interface';
 import type { BenchmarkResponse } from './responses/benchmark-response.interface';
@@ -114,9 +112,7 @@ export {
   ActivityResponse,
   AdminData,
   AdminJobs,
-  AdminMarketData,
   AdminMarketDataDetails,
-  AdminMarketDataItem,
   AdminUser,
   AdminUserResponse,
   AdminUsersResponse,
@@ -126,6 +122,8 @@ export {
   AssertionCredentialJSON,
   AssetClassSelectorOption,
   AssetProfileIdentifier,
+  AssetProfileItem,
+  AssetProfilesResponse,
   AssetResponse,
   AttestationCredentialJSON,
   Benchmark,